tidy code.
